+++ lib/td_dd/data_structures/loader_worker.ex
@@ -6,6 +6,7 @@ defmodule TdDd.Loader.LoaderWorker do
   use GenServer
 
   alias TdDd.Loader
+  alias TdDd.ProfilingLoader
   alias TdDd.Repo
 
   require Logger
@@ -24,6 +25,10 @@ defmodule TdDd.Loader.LoaderWorker do
       _ ->
         GenServer.call(__MODULE__, {:load, structures, fields, relations, audit, opts})
     end
+  end
+
+  def load(profiles) do
+    GenServer.cast(TdDd.Loader.LoaderWorker, {:profiles, profiles})
   end
 
   def ping(timeout \\ 5000) do
@@ -52,6 +57,28 @@ defmodule TdDd.Loader.LoaderWorker do
     {:reply, :pong, state}
   end
 
+  @impl true
+  def handle_cast({:profiles, profiles}, state) do
+    Logger.info("Bulk loading profiles")
+    {ms, res} = Timer.time(fn -> ProfilingLoader.load(profiles) end)
+
+    case res do
+      {:ok, ids} ->
+        count = Enum.count(ids)
+        Logger.info("Bulk load process completed in #{ms}ms (#{count} upserts)")
+
+      _ ->
+        Logger.warn("Bulk load failed after #{ms}")
+    end
+
+    {:noreply, state}
+  end
+
+  @impl true
+  def handle_call(:ping, _from, state) do
+    {:reply, :pong, state}
+  end
+
   defp do_load(structures, fields, relations, audit, opts \\ []) do
     Logger.info("Bulk loading data structures")
     {ms, res} = Timer.time(fn -> Loader.load(structures, fields, relations, audit, opts) end)
@@ -66,6 +93,8 @@ defmodule TdDd.Loader.LoaderWorker do
         Logger.warn("Bulk load failed after #{ms}ms (#{inspect(e)})")
         Repo.rollback(e)
     end
+
+    {:noreply, state}
   end
 
   defp post_process([]), do: :ok