+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-2089] Profiling support for structures
+- [TD-2118] Metadata as mapping in data structures search 
+
 ## [3.6.0] 2019-09-16
 
 ### Added